// src/modules/outpatient/checklist-pra-operasi/models/checklist-pra-operasi-models.ts
export interface IChecklistPraOperasiForm {
    T: string;
    Bb: string;
    Rr: string;
    Tb: string;
    Td: string;
    Ekg: string;
    Sat: string;
    Nadi: string;
    Hamil: string;
    Lensa: string;
    Puasa: string;
    Alergi: string;
    Lokasi: string;
    Makula: string;
    Rambut: string;
    Implant: string;
    Tanggal: string;
    Anestesi: string;
    Biometri: string;
    Gliserin: string;
    Kosmetik: string;
    Usg_Mata: string;
    Jam_Mulai: string;
    Needle_No: string;
    Perhiasan: string;
    Radiologi: string;
    Gigi_Palsu: string;
    ID_Petugas: string;
    Pembedahan: string;
    Updated_At: string;
    Updated_By: string;
    Foto_Fundus: string;
    Cairan_Masuk: string;
    Jenis_Cairan: string;
    Jenis_Pasien: string;
    Dokter_Anestesi_Id: string;
    Dokter_Anestesi_Nama: string
    Laboratorium: string;
    Nama_Petugas: string;
    Pre_Medikasi: string;
    Resiko_Jatuh: string;
    Gelang_Alergi: string;
    Kandung_Kemih: string;
    Rhesus_Fektor: string;
    Golongan_Darah: string;
    Inform_Consent: string;
    Infus_Dipasang: string;
    Catatan_Perawat: string;
    Gelang_Pengenal: string;
    Persiapan_Darah: string;
    Izin_Sterilisasi: string;
    Kondisi_Kulit_Id: string;
    Puasa_Keterangan: string;
    Alergi_Keterangan: string;
    Nama_Kepala_Bedah: string;
    Pemeriksaan_Lainnya: string;
    Inform_Consent_Bedah: string;
    Nama_Perawat_Penerima: string;
    Nama_Perawat_Pengantar: string;
    Tanda_Tangan_Kepala_Bedah: string;
    Persiapan_Darah_Keterangan: string;
    ID_Tanda_Tangan_Kepala_Bedah: string;
    Tanda_Tangan_Perawat_Penerima: string;
    Tanda_Tangan_Perawat_Pengantar: string;
    ID_Tanda_Tangan_Perawat_Penerima: string;
    ID_Tanda_Tangan_Perawat_Pengantar: string;
}
// ...
export interface IChecklistPraOperasiModel extends IDataModel {
    form: IChecklistPraOperasiForm;
    dokter: Array<IDoctorModel>;
    asesmen: IAssessmentUgdFormModel;
    ews: INursingEarlyWarning;
    tanda_vital: IPerioperativeNursingRecordsFormModel;
}

export class ChecklistPraOperasiModel extends DataModel {
    form: IChecklistPraOperasiForm;
    dokter: Array<IDoctorModel>;
    asesmen?: IAssessmentUgdFormModel;
    ews?: INursingEarlyWarning;
    tanda_vital: IPerioperativeNursingRecordsFormModel;

    constructor(req: IChecklistPraOperasiModel) {
      super(req)
      this.form = req.form;
      this.dokter = (Array.isArray(req.dokter)) ? req.dokter.map((a) => new DoctorModel(a)) : [];
      this.asesmen = req.asesmen;
      this.ews = req.ews;
      this.tanda_vital = req.tanda_vital;
    }
}